// types/Cart.tsx
export interface Cart {
  product_id: number;
  product_number: string;
  name: string;
  image: string;
  flavor: string;
  price: number;
  product_quantity: number;
  availability: string;
  cart_id: number;
  cart_quantity: number;
  favorite_id: null;
  in_cart: number;
  in_favorite: number;
}